Reduce ints to bytes

+++ Board.java
@@ -2,20 +2,20 @@ import java.util.Random;
 
 public class Board {
 	
-	public int[] backRow;
-	public int[][] whiteLocations;
-	public int[][] blackLocations;
-	public int[][] friendly;
-	public int[][] opposing;
+	public byte[] backRow;
+	public byte[][] whiteLocations;
+	public byte[][] blackLocations;
+	public byte[][] friendly;
+	public byte[][] opposing;
 	
-	/* 1 = Rook, 2 = Knight, 3 = Bishop, 5 = King,
-	 * 4 = Queen, 6 = White Pawn, 7 = Black Pawn, 0 = Empty
+	/* 1 = Rook, 2 = Knight, 3 = Bishop, 4 = Queen,
+	 * 5 = King, 6 = White Pawn, 7 = Black Pawn, 0 = Empty
 	 */
 	
 	public Board() {
-		backRow = new int[]{1, 1, 2, 2, 3, 3, 4, 5};
-		whiteLocations = new int[8][8];
-		blackLocations = new int[8][8];
+		backRow = new byte[]{1, 1, 2, 2, 3, 3, 4, 5};
+		whiteLocations = new byte[8][8];
+		blackLocations = new byte[8][8];
 		randomizeBackRow();
 		setupFrontRow();
 		friendly = whiteLocations;
@@ -23,7 +23,8 @@ public class Board {
 	}
 	
 	private void randomizeBackRow() {
-		for (int i = 0; i < backRow.length; i++) {
+		// nested for loops to make random back row
+		for (byte i = 0; i < backRow.length; i++) {
 			Random rand = new Random();
 			int randNum = rand.nextInt(backRow.length);
 			while (whiteLocations[backRow.length-1][randNum] != 0) {
@@ -35,13 +36,15 @@ public class Board {
 	}
 	
 	private void setupFrontRow() {
-		for (int i = 0; i < backRow.length; i++) {
+		// Create a row of pawns in front of each color
+		for (byte i = 0; i < backRow.length; i++) {
 			whiteLocations[backRow.length-2][i] = 6;
 			blackLocations[1][i] = 7;
 		}
 	}
 	
-	public void changeBoard(int newX, int newY, int oldX, int oldY) {
+	public void changeBoard(byte newX, byte newY, byte oldX, byte oldY) {
+		// move the piece
 		if (whiteLocations[oldX][oldY] != 0) {
 			whiteLocations[newX][newY] = whiteLocations[oldX][oldY];
 			whiteLocations[oldX][oldY] = 0;
@@ -52,6 +55,7 @@ public class Board {
 			whiteLocations[newX][newY] = 0;
 		}
 		
+		// switch turns
 		if (friendly == whiteLocations) {
 			friendly = blackLocations;
 			opposing = whiteLocations;
